Also listen to events in command (and not just in editor provider)
