+++ src/YarnEditorProvider.ts
@@ -6,12 +6,14 @@ import {
   WebviewPanel,
   Disposable,
   window,
-  workspace,
 } from "vscode";
 
 import YarnEditorPanel from "./YarnEditorWebviewPanel";
-import { writeFile } from "fs";
+import YarnEditorMessageListener from "./YarnEditorMessageListener";
 
+/**
+ * This is a custom text provider that will open up `.yarn` files in the Yarn Editor.
+ */
 export class YarnEditorProvider implements CustomTextEditorProvider {
   /**
    * This is used to trigger calling this on certain file types.
@@ -54,55 +56,10 @@ export class YarnEditorProvider implements CustomTextEditorProvider {
       enableScripts: true,
     };
 
-    webviewPanel.webview.onDidReceiveMessage((message) => {
-      switch (message.command) {
-        // this gets sent by "trySaveCurrent" in YarnEditor's data.js
-        case "save":
-          writeFile(document.fileName, message.data, (e) => {
-            if (e) {
-              window.showErrorMessage(`Error writing ${document.fileName}\n${e.message}`);
-            } else {
-              window.showInformationMessage(`Saved ${document.fileName}`);
-            }
-          });
-          break;
+    // this will listen for events from the editor
+    YarnEditorMessageListener(webviewPanel, document);
 
-        // we override `window.alert` to send messages here in YarnEditorWebviewPanel
-        case "alert":
-          window.showWarningMessage(message.data);
-          break;
-
-        case "changeSetting":
-          // triggered whenever the user changes a setting in the YarnEditor's settings window
-          const { option, newValue } = message.data;
-
-          const yarnSpinnerSettings = workspace.getConfiguration("yarnSpinner");
-
-          // if "overrideDarkThemeNightMode" is not true, we ignore saving "nightModeEnabled"
-          if (yarnSpinnerSettings.get("overrideDarkThemeNightMode") === false && option === "nightModeEnabled") {
-            break;
-          }
-
-          // these always come back as strings, but sometimes we want a boolean or a number...
-          // we try to parse it as JSON to get the proper data type and if that fails, we just use the value (since it's probably a string)
-          let parsedValue;
-          try {
-            parsedValue = JSON.parse(newValue);
-          } catch (e) {
-            parsedValue = newValue;
-          }
-
-          // if the value has changed, save it
-          if (yarnSpinnerSettings.get(option) !== parsedValue) {
-            yarnSpinnerSettings.update(option, parsedValue);          
-          }
-
-          break;
-        default:
-          break;
-      }
-    });
-
+    // this actually renders the editor
     YarnEditorPanel(webviewPanel, this.context.extensionPath, document);
   }
 }
